Stop assn-4
The linker files were all made for 32 bit systems, addded -m32 flag to
gcc to run it but it failed despite that as the the url.o file only
worked with http and the rss feeds are all https now. Note to my future
self, incase I ever resume the work here.